Revert "wire type"
This reverts commit 77aa4fefe0b4a09e012b01b15b8c8fc6beb5d2ab.
