+++ canbench-bin/tests/expected/reports_repeated_scope_in_existing_benchmark.txt
@@ -2,7 +2,7 @@
 
 Benchmark: bench_repeated_scope_exists
   total:
-    instructions: 16.46 K (regressed from 0)
+    instructions: 16.49 K (regressed from 0)
     heap_increase: 0 pages (no change)
     stable_memory_increase: 0 pages (no change)
 
@@ -18,7 +18,7 @@ Summary:
   instructions:
     status:   Regressions detected 🔴
     counts:   [total 1 | regressed 1 | improved 0 | new 0 | unchanged 0]
-    change:   [max +16.46K | p75 +16.46K | median +16.46K | p25 +16.46K | min +16.46K]
+    change:   [max +16.49K | p75 +16.49K | median +16.49K | p25 +16.49K | min +16.49K]
     change %: [max +inf% | p75 +inf% | median +inf% | p25 +inf% | min +inf%]
 
   heap_increase:
@@ -38,7 +38,7 @@ Summary:
 Only significant changes:
 | status | name                                 | calls |    ins |   ins Δ% | HI |    HI Δ% | SMI |  SMI Δ% |
 |--------|--------------------------------------|-------|--------|----------|----|----------|-----|---------|
-|   +    | bench_repeated_scope_exists          |       | 16.46K |    +inf% |  0 |    0.00% |   0 |   0.00% |
+|   +    | bench_repeated_scope_exists          |       | 16.49K |    +inf% |  0 |    0.00% |   0 |   0.00% |
 |  +/-   | bench_repeated_scope_exists::scope_1 |    10 |  8.73K | +991.75% |  0 | -100.00% |   0 |   0.00% |
 
 ins = instructions, HI = heap_increase, SMI = stable_memory_increase, Δ% = percent change
