+++ canbench-bin/tests/expected/reports_repeated_scope_in_new_benchmark.txt
@@ -2,7 +2,7 @@
 
 Benchmark: bench_repeated_scope_new (new)
   total:
-    instructions: 16.46 K (new)
+    instructions: 16.49 K (new)
     heap_increase: 0 pages (new)
     stable_memory_increase: 0 pages (new)
 
@@ -38,7 +38,7 @@ Summary:
 Only significant changes:
 | status | name                              | calls |    ins |  ins Δ% | HI |  HI Δ% | SMI |  SMI Δ% |
 |--------|-----------------------------------|-------|--------|---------|----|--------|-----|---------|
-|  new   | bench_repeated_scope_new          |       | 16.46K |         |  0 |        |   0 |         |
+|  new   | bench_repeated_scope_new          |       | 16.49K |         |  0 |        |   0 |         |
 |  new   | bench_repeated_scope_new::scope_1 |    10 |  8.73K |         |  0 |        |   0 |         |
 
 ins = instructions, HI = heap_increase, SMI = stable_memory_increase, Δ% = percent change
