+++ canbench-bin/tests/expected/reports_scopes_in_existing_benchmark.txt
@@ -2,7 +2,7 @@
 
 Benchmark: bench_scope_exists
   total:
-    instructions: 4134 (regressed from 0)
+    instructions: 4140 (regressed from 0)
     heap_increase: 0 pages (no change)
     stable_memory_increase: 0 pages (no change)
 
@@ -24,7 +24,7 @@ Summary:
   instructions:
     status:   Regressions detected 🔴
     counts:   [total 1 | regressed 1 | improved 0 | new 0 | unchanged 0]
-    change:   [max +4.13K | p75 +4.13K | median +4.13K | p25 +4.13K | min +4.13K]
+    change:   [max +4.14K | p75 +4.14K | median +4.14K | p25 +4.14K | min +4.14K]
     change %: [max +inf% | p75 +inf% | median +inf% | p25 +inf% | min +inf%]
 
   heap_increase:
@@ -44,7 +44,7 @@ Summary:
 Only significant changes:
 | status | name                        | calls |   ins |  ins Δ% | HI |    HI Δ% | SMI |  SMI Δ% |
 |--------|-----------------------------|-------|-------|---------|----|----------|-----|---------|
-|   +    | bench_scope_exists          |       | 4.13K |   +inf% |  0 |    0.00% |   0 |   0.00% |
+|   +    | bench_scope_exists          |       | 4.14K |   +inf% |  0 |    0.00% |   0 |   0.00% |
 |  +/-   | bench_scope_exists::scope_1 |     1 | 1.11K | +38.88% |  0 | -100.00% |   0 |   0.00% |
 |  new   | bench_scope_exists::scope_2 |     1 |   847 |         |  0 |          |   0 |         |
