+++ canbench-bin/tests/expected/reports_scopes_in_new_benchmark.txt
@@ -2,7 +2,7 @@
 
 Benchmark: bench_scope_new (new)
   total:
-    instructions: 4134 (new)
+    instructions: 4140 (new)
     heap_increase: 0 pages (new)
     stable_memory_increase: 0 pages (new)
 
@@ -44,7 +44,7 @@ Summary:
 Only significant changes:
 | status | name                     | calls |   ins |  ins Δ% | HI |  HI Δ% | SMI |  SMI Δ% |
 |--------|--------------------------|-------|-------|---------|----|--------|-----|---------|
-|  new   | bench_scope_new          |       | 4.13K |         |  0 |        |   0 |         |
+|  new   | bench_scope_new          |       | 4.14K |         |  0 |        |   0 |         |
 |  new   | bench_scope_new::scope_1 |     1 | 1.11K |         |  0 |        |   0 |         |
 |  new   | bench_scope_new::scope_2 |     1 |   847 |         |  0 |        |   0 |         |
