+++ examples/btreemap_vs_hashmap/canbench_results.yml
@@ -9,7 +9,7 @@ benches:
   pre_upgrade_bench:
     total:
       calls: 1
-      instructions: 486879758
+      instructions: 486879776
       heap_increase: 519
       stable_memory_increase: 184
     scopes:
